src/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
autopep8==1.5.7
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.0.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
Django==3.2.4
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-currentuser==0.5.3
django-filter==21.1
django-jazzmin==2.5.0
django-js-asset==1.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
future==0.18.2
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==6.0
reportlab==3.5.68
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
zope.interface==5.4.0
